Convert profiling modules to packages with package scoped data fixtures
