+++ tests/profiling/splash/test_profiling_splash.py
